requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
Pillow
python-multipart
h5py
